fix: renamed component & restructured component
